batchCalculation: Allow to re-use a previous calculation
fixes #580

Update the type of the csvFile, to either be an upload or a server file
from a specific jobId.

Add a socket route to fetch a job's parameters to replay it. It gets the
csvFields from the job's input file, as well as the transit routing
parameters and demand data.

The frontend forms are updated to account for this new calculation
initialization path:

* The initial demand and routing data is optionally passed as props to
  the calculation form
* The upload button only appears if the file type is an upload and the
  demand data is valid.
* Any of the parameter from the previous job is editable in the current
  configuration, including the uploaded file. If the user decides to use
another file, the 'server' type is changed to 'upload' and the file will
have to be uploaded before going to the next step.