batchCalculation: Change type of the demandAttributes
The parameters was previously called batchRoutingAttributes, which is
confusing with transitRoutingAttributes. Also, 2 distinct types were
used, one from the chaire-lib-common api called
TransitBatchRoutingAttributes, the other from the transitDemand data,
which contained the same fields, but optional instead of required.

The parameter is renamed to demandAttributes.

The type from chaire-lib-common is thus updated to represent the
demandAttributes, with the same mandatory fields. The
TransitDemandFromCsv attributes are built from the former, but as
optional fields.