jobs: Get files with their actual name in title
This is more informative to the user, especially for input files, which
now can have different names. The title of the files now contain both
the text describing the file, and the file name.

Localized strings can use `{{filename}}` to display the actual file
name. Input files now do so for batch routing and batch accessibility
map jobs.